models update, base model fixes
