+++ users/models/__init__.py
@@ -1,3 +1,3 @@
 from .user import User
-
-# from .other import AddressRegister
+from .other import AddressRegister
+from .base_model import BaseModel
