+++ users/models/user.py
@@ -1,8 +1,6 @@
 import datetime
 from django.db import models
 from django.contrib.auth.models import AbstractUser
-# from setup.models import BaseModel
-
 from rest_framework_simplejwt.tokens import RefreshToken
 
 
@@ -28,6 +26,30 @@ class User(AbstractUser):
     profile_picture = models.FileField(upload_to='profile/', blank=True, null=True, verbose_name='Profile Picture')
 
     is_suspended = models.BooleanField(default=False)
+
+    ########################################
+    # BaseModel
+    ########################################
+    created_by = models.ForeignKey('User', related_name='%(class)s_created_by',
+                                   on_delete=models.SET_NULL,
+                                   null=True, blank=True,
+                                   verbose_name='Created By')
+    created_at = models.DateTimeField(auto_now=True, auto_now_add=False, verbose_name='Created At')
+
+    updated_by = models.ForeignKey('User', related_name='%(class)s_updated_by',
+                                   on_delete=models.SET_NULL,
+                                   null=True, blank=True,
+                                   verbose_name='Updated By')
+    updated_at = models.DateTimeField(auto_now=False, auto_now_add=True, verbose_name='Updated At')
+
+    deleted = models.BooleanField(default=False, verbose_name='Deleted')
+    deleted_at = models.DateTimeField(auto_now=True, auto_now_add=False, verbose_name='Deleted At')
+    deleted_by = models.ForeignKey('User', related_name='%(class)s_deleted_by',
+                                   on_delete=models.SET_NULL,
+                                   null=True, blank=True,
+                                   verbose_name='Deleted By')
+
+    ########################################
 
     class Meta(AbstractUser.Meta):
         pass
